requirements.txt
# --- Web API Server (FastAPI) ---
fastapi
uvicorn[standard]
python-multipart
python-dotenv

# --- gRPC Service ---
grpcio
grpcio-tools
grpcio-reflection

# --- Diagram & Image Processing (繪圖與影像) ---
diagrams
graphviz
Pillow

# --- Crawler & Scraper (爬蟲與排程) ---
requests
beautifulsoup4
facebook-scraper
googlesearch-python
feedparser
schedule
apscheduler
duckduckgo-search
fake-useragent

# --- Process Management & Dev Tools (管理與開發) ---
honcho
watchdog
sqlalchemy